Moved monitored trip tests into separate class
