+++ src/test/java/org/opentripplanner/middleware/persistence/PersistenceTest.java
@@ -2,12 +2,10 @@ package org.opentripplanner.middleware.persistence;
 
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.middleware.OtpMiddlewareTest;
-import org.opentripplanner.middleware.models.MonitoredTrip;
 import org.opentripplanner.middleware.models.OtpUser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.opentripplanner.middleware.persistence.PersistenceUtil.createMonitoredTrip;
 import static org.opentripplanner.middleware.persistence.PersistenceUtil.createUser;
 
 /**
@@ -49,15 +47,4 @@ public class PersistenceTest extends OtpMiddlewareTest {
         OtpUser user = Persistence.otpUsers.getById(userToDelete.id);
         assertNull(user, "Deleted User should no longer exist in database (should return as null).");
     }
-
-    @Test
-    public void canCreateMonitoredTrip() {
-        String userId = "123456";
-        MonitoredTrip monitoredTrip = createMonitoredTrip(userId);
-        MonitoredTrip retrieved = Persistence.monitoredTrip.getById(monitoredTrip.id);
-        assertEquals(monitoredTrip.id, retrieved.id, "Found monitored trip ID should equal inserted ID.");
-        // tidy up
-        Persistence.monitoredTrip.removeById(monitoredTrip.id);
-    }
-
 }
